% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condlSample.R
\name{condlSample}
\alias{condlSample}
\alias{condlSample.rcgam}
\alias{condlSample.rclm}
\title{Generate samples from conditional distribution.}
\usage{
condlSample(object, ...)

\method{condlSample}{rcgam}(object, newdata, flowcol = "flow",
  flow.units = "ft3/s", quantile, ...)

\method{condlSample}{rclm}(object, newdata, flowcol = "flow",
  flow.units = "ft3/s", quantile, ...)
}
\arguments{
\item{object}{a model object with a `predict()` method}

\item{...}{arguments to be passed to individual methods. These should always include :
`newdata` - to be passed to `predict()` function
`quantile` - either 'random' for stochastic sampling from conditional distribution or
a numeric value on (0, 1) specifying the quantile to return}

\item{newdata}{a data.frame containing precictor variables to use for prediction}

\item{object}{An rclm or rcgam object to use for predicting}

\item{retransform}{Should the predictions be returned as concentrations? (defaults to TRUE)}

\item{...}{Arguments passed to `predict.lm` or predict.gam function call}

\item{smear}{Use Smearing estimator to correct transformation bias?}

\item{object}{An rclm or rcgam object to use for predicting}

\item{newdata}{a data.frame containing precictor variables to use for prediction}

\item{retransform}{Should the predictions be returned as concentrations? (defaults to TRUE)}

\item{...}{Arguments passed to `predict.lm` or predict.gam function call}

\item{smear}{Use Smearing estimator to correct transformation bias?}
}
\value{
Numeric vector containing conditional random sample (if `quantile = 'random'`)
or conditional quantiles from condition distribution defined by `object` and `newdata`
seealso \code{\link{condlSample.lm}}
}
\description{
Generate samples from conditional distribution.

Generate samples from a conditional distribution obtained from a model prediction

Generate samples from a conditional distribution obtained from a model prediction
}
\section{Methods (by class)}{
\itemize{
\item \code{rcgam}: Method for rcgam objects

\item \code{rclm}: Method for rclm objects
}}

